requirements.txt
Django==4.0
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt==5.1.0
Pillow==7.0.0
drf-yasg
